exoplanet/backend/app/requirements.txt
# Deep Learning Framework
# PyTorch with MPS support for Apple Silicon GPU acceleration
torch>=2.0.0
torchvision>=0.15.0

# CLIP for planetary image gatekeeper
open-clip-torch>=2.20.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Data Handling
h5py>=3.8.0
matplotlib>=3.6.0
mat73>=0.60  # For MATLAB v7.3 format files

# Machine Learning
scikit-learn>=1.3.0

# Progress Tracking
tqdm>=4.65.0

# Logging
colorlog>=6.7.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=9.0.0

# Visualization & Analysis
seaborn>=0.12.0

# Jupyter Notebook Support
notebook>=6.5.0
ipywidgets>=8.0.0

# Development Tools (Optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pytest>=7.0.0

joblib>=0.14
matplotlib>=3.1
numpy>=1.18
scipy>=1.4
spectral>=0.19

# optional optimizations
Bottleneck>=1.3
mat73>=0.46

# API Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Model Download
gdown>=4.6.0
